Move perimetre editing form under editing item
