+++ components/suivi-form/perimetres/index.js
@@ -1,114 +1,23 @@
-import {useState, useCallback, useEffect, useMemo} from 'react'
+import {useState} from 'react'
 import PropTypes from 'prop-types'
-import {debounce} from 'lodash-es'
-
-import {getPerimetersByName, getPerimetersByCode} from '@/lib/decoupage-administratif-api.js'
 
 import colors from '@/styles/colors.js'
 
-import AutocompleteInput from '@/components/autocomplete-input.js'
 import Perimetre from '@/components/suivi-form/perimetres/perimetre.js'
-import SelectInput from '@/components/select-input.js'
+import PerimetreForm from '@/components/suivi-form/perimetres/perimetre-form.js'
 import Button from '@/components/button.js'
 
-const TYPES = [
-  {label: 'EPCI', value: 'epci'},
-  {label: 'Commune', value: 'commune'},
-  {label: 'Département', value: 'departement'}
-]
-
-const renderItem = (item, isHighlighted) => {
-  const {nom, code} = item
-
-  return (
-    <div key={code} className='item fr-px-1w fr-py-2w'>
-      {nom} - {code}
-
-      <style jsx>{`
-        .item {
-          background: ${isHighlighted ? colors.blueHover : 'white'};
-          color: ${isHighlighted ? 'white' : colors.darkgrey};
-        }
-
-        .ape {
-          font-weight: bold;
-        }
-      `}</style>
-    </div>
-  )
-}
-
 const Perimetres = ({perimetres, hasMissingData, handlePerimetres, onRequiredFormOpen}) => {
-  const [isFormOpen, setIsFormOpen] = useState()
-  const [hasMissingInput, setHasMissingInput] = useState(false)
-  const [isLoading, setIsLoading] = useState(false)
-  const [errorMessage, setErrorMessage] = useState()
-  const [code, setCode] = useState('')
-  const [nom, setNom] = useState('')
-
-  const [type, setType] = useState('')
-
-  const [foundPerimetres, setFoundPerimetres] = useState([])
+  const [isAdding, setIsAdding] = useState(false)
+  const [isEditing, setIsEditing] = useState(false)
   const [updatingPerimetreIndex, setUpdatingPerimetreIndex] = useState()
-  const [updatingPerimetreCode, setUpdatingPerimetreCode] = useState()
 
-  const isUpdating = useMemo(() => (updatingPerimetreIndex || updatingPerimetreIndex === 0) && true, [updatingPerimetreIndex])
-
-  const handleErrors = useCallback(input => {
-    if (!input && hasMissingInput) {
-      return 'Ce champs est requis'
-    }
-  }, [hasMissingInput])
-
-  const onReset = () => {
-    setNom('')
-    setCode('')
-    setType('')
-    setUpdatingPerimetreIndex(null)
-    setIsFormOpen(false)
-    onRequiredFormOpen(false)
-    setHasMissingInput(false)
-    setUpdatingPerimetreCode(null)
-    setErrorMessage(null)
-  }
-
-  const onAdd = () => {
-    if (type && code) {
-      if (perimetres.includes(`${type}:${code}`)) {
-        setErrorMessage('Ce périmètre est déjà présent.')
-      } else {
-        handlePerimetres([...perimetres, `${type}:${code}`])
-
-        onReset()
-      }
-    } else {
-      setHasMissingInput(true)
-    }
-  }
-
-  const onUpdate = () => {
-    if (type && code) {
-      if (perimetres.includes(`${type}:${code}`) && updatingPerimetreCode !== code) {
-        setErrorMessage('Ce périmètre est déjà présent.')
-      } else {
-        handlePerimetres([...perimetres].map((perimetre, idx) => {
-          if (idx === updatingPerimetreIndex) {
-            perimetre = `${type}:${code}`
-          }
-
-          return perimetre
-        }))
-
-        onReset()
-      }
-    } else {
-      setHasMissingInput(true)
-    }
-  }
+  const isFormOpen = isAdding || isEditing
 
   const onDelete = index => {
     handlePerimetres(current => current.filter((_, i) => index !== i))
-    onReset()
+    setIsAdding(false)
+    setIsEditing(false)
   }
 
   const perimetreAsObject = perimetre => {
@@ -117,73 +26,6 @@ const Perimetres = ({perimetres, hasMissingData, handlePerimetres, onRequiredFor
     const perimetreCode = perimetreAsArray[1]
 
     return {perimetreType, perimetreCode}
-  }
-
-  useEffect(() => {
-    // Switch to perimetre update form
-    if (updatingPerimetreIndex || updatingPerimetreIndex === 0) {
-      const {perimetreCode, perimetreType} = perimetreAsObject(perimetres[updatingPerimetreIndex])
-      try {
-        const switchToUpdateForm = async () => {
-          const perimetre = await getPerimetersByCode(perimetreCode, perimetreType)
-          setNom(perimetre.nom)
-          setCode(perimetre.code)
-          setType(perimetreType)
-          setUpdatingPerimetreCode(perimetre.code)
-          setIsFormOpen(true)
-          onRequiredFormOpen(true)
-        }
-
-        switchToUpdateForm()
-      } catch {
-        console.log('Impossible de retrouver le périmètre')
-      }
-    }
-  }, [perimetres, updatingPerimetreIndex, onRequiredFormOpen])
-
-  const fetchPerimetres = useCallback(debounce(async (nom, type, signal) => {
-    setIsLoading(true)
-
-    try {
-      const perimetres = await getPerimetersByName(nom, type, signal)
-      setFoundPerimetres(perimetres)
-    } catch {
-      if (!signal.aborted) {
-        setFoundPerimetres([])
-      }
-    }
-
-    setIsLoading(false)
-  }, 300), [setFoundPerimetres, setIsLoading])
-
-  useEffect(() => {
-    if (!nom || nom.length < 3) {
-      return
-    }
-
-    const ac = new AbortController()
-    fetchPerimetres(nom, type, ac.signal)
-
-    return () => {
-      ac.abort()
-    }
-  }, [nom, type, fetchPerimetres])
-
-  const handleChange = value => {
-    setType(value)
-
-    if (value !== type) {
-      setNom('')
-      setCode('')
-      setFoundPerimetres([])
-    }
-  }
-
-  const handleSelect = item => {
-    const foundPerimetreName = foundPerimetres.find(result => result.code === item).nom
-
-    setCode(item)
-    setNom(foundPerimetreName)
   }
 
   return (
@@ -197,19 +39,55 @@ const Perimetres = ({perimetres, hasMissingData, handlePerimetres, onRequiredFor
 
       <div className='fr-grid-row fr-mt-2w'>
         {perimetres.map((perimetre, idx) => (
-          <Perimetre
-            key={perimetre}
-            perimetre={perimetre}
-            perimetreAsObject={perimetreAsObject(perimetre)}
-            handleUpdate={() => setUpdatingPerimetreIndex(idx)}
-            handleDelete={() => onDelete(idx)}
-            onReset={onReset}
-          />
+          <div key={perimetre} className={updatingPerimetreIndex === idx ? 'fr-col-12' : ''}>
+            <Perimetre
+              perimetre={perimetre}
+              perimetreAsObject={perimetreAsObject(perimetre)}
+              isFormOpen={isFormOpen}
+              handleUpdate={() => {
+                setIsEditing(true)
+                setUpdatingPerimetreIndex(idx)
+              }}
+              handleDelete={() => onDelete(idx)}
+            />
+
+            {updatingPerimetreIndex === idx && (
+              <div>
+                <PerimetreForm
+                  perimetreAsObject={perimetreAsObject}
+                  updatingPerimetreIdx={updatingPerimetreIndex}
+                  handleUpdatingPerimetreIdx={setUpdatingPerimetreIndex}
+                  handleAdding={setIsAdding}
+                  handleEditing={setIsEditing}
+                  perimetres={perimetres}
+                  handlePerimetres={handlePerimetres}
+                  isEditing={isEditing}
+                  onRequiredFormOpen={onRequiredFormOpen}
+                />
+
+                <div className='edit-separator fr-my-3w' />
+              </div>
+            )}
+          </div>
         )
         )}
       </div>
 
-      {!isFormOpen && (
+      {isAdding && (
+        <PerimetreForm
+          perimetreAsObject={perimetreAsObject}
+          updatingPerimetreIdx={updatingPerimetreIndex}
+          handleUpdatingPerimetreIdx={setUpdatingPerimetreIndex}
+          handleAdding={setIsAdding}
+          handleEditing={setIsEditing}
+          perimetres={perimetres}
+          handlePerimetres={handlePerimetres}
+          isEditing={isEditing}
+          onRequiredFormOpen={onRequiredFormOpen}
+        />
+      )}
+
+      {!isAdding && !isEditing && (
         <div className='fr-mt-3w'>
           <Button
             label='Ajouter un périmètre'
@@ -217,7 +95,7 @@ const Perimetres = ({perimetres, hasMissingData, handlePerimetres, onRequiredFor
             iconSide='left'
             onClick={() => {
               onRequiredFormOpen(true)
-              setIsFormOpen(true)
+              setIsAdding(true)
             }}
           >
             Ajouter un périmètre
@@ -225,67 +103,13 @@ const Perimetres = ({perimetres, hasMissingData, handlePerimetres, onRequiredFor
         </div>
       )}
 
-      {isFormOpen && (
-        <div>
-          <div className='fr-grid-row fr-my-5w'>
-            <div className='fr-col-12 fr-col-md-6 fr-mb-3w'>
-              <SelectInput
-                isRequired
-                label='Type'
-                value={type}
-                options={TYPES}
-                ariaLabel='types de territoire'
-                description='Types de territoire'
-                errorMessage={handleErrors(type)}
-                onValueChange={e => handleChange(e.target.value)}
-              />
-            </div>
-
-            {type && (
-              <div className='fr-col-12 fr-col-md-6 fr-pl-md-3w'>
-                <AutocompleteInput
-                  isRequired
-                  label='Nom'
-                  value={nom}
-                  description='Nom du territoire *'
-                  ariaLabel='rechercher le nom du territoire'
-                  errorMessage={handleErrors(nom)}
-                  results={foundPerimetres}
-                  customItem={renderItem}
-                  isLoading={isLoading}
-                  getItemValue={item => item.code}
-                  onValueChange={e => setNom(e.target.value)}
-                  onSelectValue={item => handleSelect(item)}
-                />
-              </div>
-            )}
-          </div>
-
-          <div className='fr-grid-row'>
-            <Button
-              label='Valider l’ajout du périmètre'
-              icon='checkbox-circle-fill'
-              onClick={() => isUpdating ? onUpdate() : onAdd()}
-            >
-              Valider
-            </Button>
-            <div className='fr-pl-3w'>
-              <Button
-                label='Annuler l’ajout du périmètre'
-                buttonStyle='tertiary'
-                onClick={onReset}
-              >
-                Annuler
-              </Button>
-            </div>
-          </div>
-          {errorMessage && <p id='text-input-error-desc-error' className='fr-error-text'>{errorMessage}</p>}
-        </div>
-      )}
-
       <style jsx>{`
-        .separator {
-          border-top: 3px solid ${colors.grey850};
+        .editing-perimetre {
+          flex: 1;
+        }
+
+        .edit-separator {
+          border-top: 2px solid ${colors.blueFrance850};
         }
       `}</style>
     </div>
@@ -302,5 +126,4 @@ Perimetres.propTypes = {
 Perimetres.defaultProps = {
   hasMissingData: false
 }
-
 export default Perimetres
